MemSink to datasource
