+++ datafusion-examples/examples/custom_datasource.rs
@@ -37,6 +37,7 @@ use datafusion::physical_plan::{
     PlanProperties, SendableRecordBatchStream,
 };
 use datafusion::prelude::*;
+use datafusion_datasource::memory::PartitionData;
 
 use datafusion::catalog::Session;
 use tokio::time::timeout;
